\name{AngristLavy}
\docType{data}
\alias{AngristLavy}
\title{Angrist and Lavy (1999) Maimonides' Rule Data Set}

\description{
  Angrist and Lavy (1999) data set used by Rosenbaum (2010) to
  demonstrate his instrumental variable sensitivity analysis.
}
\usage{AngristLavy}
\format{A \code{data.frame} containing 86 matched classroom pairs (172
  rows and 9 columns).}
\source{
  Angrist, Joshua and Lavy, Victor (1999). ``Using Maimonides' Rule to
  Estimate the Effect of Class Size on Scholastic
  Achievement''. \emph{Quarterly Journal of Economics} 114, 533--575.

  Rosenbaum, Paul R. (2010). \emph{Design of Observational Studies}.
  Springer-Verlag.
}
\references{
  Rosenbaum, Paul R. (2010). \emph{Design of Observational Studies}.
  Springer-Verlag.
}
\keyword{datasets}
