% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.rbounds}
\alias{print.rbounds}
\title{A simple print method for rbounds objects}
\usage{
\method{print}{rbounds}(x, ...)
}
\arguments{
\item{x}{A model fit by \code{pidoutcomes}}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
A simple print method for rbounds objects
}
\examples{
\dontrun{
N <- 1000
x <- rnorm(N)
e <- rnorm(N)
y <- as.numeric(2*x + e > 0)
z <- rbinom(N, 1, .75)
y_obs <- z*y
df <- data.frame(y_obs, x, z)
m1 <- pidoutcomes(y_obs ~ x, z, df)
m1
}

}
