% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.rbounds_ate}
\alias{print.rbounds_ate}
\title{Print method for rbounds_ate objects}
\usage{
\method{print}{rbounds_ate}(x, ...)
}
\arguments{
\item{x}{An object of class 'rbounds_ate' estimated by \code{ate_bounds}}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for rbounds_ate objects
}
