\name{print.rbounds}
\alias{print.rbounds}
\alias{summary.rbounds}
\title{Summarize output from rbounds object}
\description{Summary method for \code{rbounds} object.}
\usage{\method{print}{rbounds}(x, ...)}

\arguments{
  \item{x}{
    An object of class \code{rbounds} as produced by
    \code{\link{binarysens}}, \code{\link{hlsens}}, or
    \code{\link{psens}}.
  }
  \item{...}{Any additional arguments.}
}

\author{
  Jason W. Morgan, Ohio State University, \email{morgan.746@osu.edu}
}

\seealso{
  Also see \code{\link{binarysens}}, \code{\link{psens}}, and
  \code{\link{hlsens}}.
}
