% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{print.rbounds}
\alias{print.rbounds}
\title{A simple print method for rbounds objects}
\usage{
\method{print}{rbounds}(m)
}
\arguments{
\item{m}{A model fit by \code{pidoutcomes}}
}
\value{
By default we print the mean upper and lower bounds and CI's
}
\description{
A simple print method for rbounds objects
}
\examples{
#' N <- 1000
x <- rnorm(N)
e <- rnorm(N)
y <- as.numeric(2*x + e > 0)
z <- rbinom(N, 1, .75)
y_obs <- z*y
df <- data.frame(y_obs, x, z)
m1 <- pidoutcomes(y_obs ~ x, z, df)
m1

}
